+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## Bugfixes
 
-* None
+* Set correct class name for Core\Api\Node\StorageServerNode to avoid warning if use Composer --optimizer-autoloader
 
 ## Features
 
-* Bugfix for Issue #286 - Version number in server software signature
+* Closed #286 - Version number in server software signature
 * Closed #294 - Session-ID structure
 * Closed #288 - Session-ID will be reused
 * Closed #292 - Annotation based configuration for servlets